fix(windowing/kwin): classify Plasma 6 clients
